% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleYhistogram.R
\name{doubleYhist}
\alias{doubleYhist}
\title{Frequency plus Relative Frequency Histogram on a Double Y Axis Scale}
\usage{
doubleYhist(dataset)
}
\arguments{
\item{dataset}{A dataset}
}
\value{
A double y scale histogram plus pie chart
}
\description{
Frequency plus Relative Frequency Histogram on a Double Y Axis Scale
}
\examples{
doubleYhist(TPRdataset)
}
