% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedhist.R
\name{groupedhist}
\alias{groupedhist}
\title{Displays VLXT, VSL2 and PONDRFIT next to each other for low, medium and high disorder categories}
\usage{
groupedhist(dataset)
}
\arguments{
\item{dataset}{A dataset containing the columns VLXT, VSL2 and PONDRFIT}
}
\value{
A side-by-side grouped histogram
}
\description{
Displays VLXT, VSL2 and PONDRFIT next to each other for low, medium and high disorder categories
}
\examples{
groupedhist(TPRdataset)
}
