% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot3dDM.R
\name{scatterplot3dDM}
\alias{scatterplot3dDM}
\title{3 Dimensional Scatterplot for VLXT, VSL2 and PONDRFIT incl Disease and/or Mutagenesis Distinction}
\usage{
scatterplot3dDM(dataset)
}
\arguments{
\item{dataset}{A dataset with columns VLXT, VSL2, PONDRFIT and DM}
}
\value{
3 dimensional scatterplot
}
\description{
3 Dimensional Scatterplot for VLXT, VSL2 and PONDRFIT incl Disease and/or Mutagenesis Distinction
}
\examples{
scatterplot3dDM(TPRdataset)
}
