% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLXTpie.R
\name{VLXTpie}
\alias{VLXTpie}
\title{Displays VLXT as pie graph}
\usage{
VLXTpie(dataset)
}
\arguments{
\item{dataset}{A dataset}
}
\value{
a pie graphs
}
\description{
Displays VLXT as pie graph
}
\examples{
VLXTpie(TPRdataset)
}
